move searchInput file and add loading state
