+++ src/components/trendingBooks/TrendingBooks.tsx
@@ -6,11 +6,13 @@ import BookCard from "./BookCard";
 
 export default function TrendingBooks() {
   const [booksToShow, setBooksToShow] = useState<Array<OpenLibraryBook>>([]);
+  const [loading, setLoading] = useState(true);
 
   const fetchTrendingBooks = async () => {
     const trendingData = await OpenLibrary.getTrendingBooks();
     const { works } = trendingData;
     setBooksToShow(works.slice(0, 20));
+    setLoading(false)
   };
 
   useEffect(() => {
@@ -21,7 +23,8 @@ export default function TrendingBooks() {
     <div>
       <span className="font-bold text-3xl">Trending Books</span>
       <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
-        {booksToShow.map((book: OpenLibraryBook) => {
+        {loading ? <span>Loading...</span> :
+          booksToShow.map((book: OpenLibraryBook) => {
           const author = Array.isArray(book.author_name)
             ? book.author_name[0]
             : book.author_name;
